requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.3.4
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bandit==1.7.5
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
coverage==7.2.7
cryptography==41.0.3
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
Django==3.2.20
django-braces==1.15.0
django-ckeditor==6.3.0
django-cors-headers==3.13.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-mptt==0.14.0
django-multiurl==1.5.0
django-oauth-toolkit==2.3.0
django-phonenumber-field-for-garpix-user==8.0.0
django-polymorphic==3.1.0
django-polymorphic-tree-for-garpix-page==2.1.1
django-rest-framework-social-oauth2==1.1.0
django-solo==2.1.0
django-tabbed-admin==1.0.4
django-tag-parser==3.2
django-uuslug==2.0.0
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.4
environs==9.5.0
eqator==2.7.0
exceptiongroup==1.1.2
fcm-django==0.3.11
filelock==3.12.2
flake8==6.1.0
flake8-polyfill==1.0.2
future==0.18.3
garpix-admin-lock==1.2.0
garpix-auth==2.3.0
garpix-menu==1.16.0
garpix-notify==5.16.0
garpix-package==2.0.1
garpix-page==2.48.1
garpix-user==3.9.0
garpix-utils==1.9.0
garpixcms==4.5.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
identify==2.5.26
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
keyring==24.2.0
kombu==5.3.1
lazy-object-proxy==1.9.0
m2r==0.2.1
mando==0.7.1
markdown-it-py==3.0.0
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
model-bakery==1.13.0
more-itertools==10.1.0
msgpack==1.0.5
nodeenv==1.8.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pbr==5.11.1
phonenumbers==8.13.18
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyfcm==1.4.9
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-telegram-bot==20.4
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
radon==6.0.1
rarfile==4.0
readme-renderer==40.0
redis==4.6.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rpds-py==0.9.2
sentry-sdk==1.29.2
service-identity==23.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.1.0
social-auth-core==4.3.0
sqlparse==0.4.4
stevedore==5.1.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
twilio==7.10.0
twine==3.4.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
unrar==0.4
uritemplate==4.1.1
urllib3==2.0.4
uWSGI==2.0.22
viberbot==1.0.12
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
zipp==3.16.2
zope.interface==6.0
